# swagger/laptop_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "laptop_service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "LaptopService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/laptop/create": {
      "post": {
        "operationId": "LaptopService_CreateLaptop",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CreateLaptopResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateLaptopRequest"
            }
          }
        ],
        "tags": [
          "LaptopService"
        ]
      }
    },
    "/v1/laptop/rate": {
      "post": {
        "operationId": "LaptopService_RateLaptop",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/RateLaptopResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of RateLaptopResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": " (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RateLaptopRequest"
            }
          }
        ],
        "tags": [
          "LaptopService"
        ]
      }
    },
    "/v1/laptop/search": {
      "get": {
        "operationId": "LaptopService_SearchLaptop",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/SearchLaptopResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of SearchLaptopResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "filter.maxPriceUsd",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "double"
          },
          {
            "name": "filter.minCpuCores",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "filter.minCpuGhz",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "double"
          },
          {
            "name": "filter.minRam.value",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "filter.minRam.unit",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "UNKNOWN",
              "BIT",
              "BYTE",
              "KILOBYTE",
              "MEGABYTE",
              "GIGABYTE",
              "TERABYTE"
            ],
            "default": "UNKNOWN"
          }
        ],
        "tags": [
          "LaptopService"
        ]
      }
    },
    "/v1/laptop/upload_image": {
      "post": {
        "operationId": "LaptopService_UploadImage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/UploadImageResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": " (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UploadImageRequest"
            }
          }
        ],
        "tags": [
          "LaptopService"
        ]
      }
    }
  },
  "definitions": {
    "CPU": {
      "type": "object",
      "properties": {
        "brand": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "numberCores": {
          "type": "integer",
          "format": "int64"
        },
        "numberThreads": {
          "type": "integer",
          "format": "int64"
        },
        "minGhz": {
          "type": "number",
          "format": "double"
        },
        "maxGhz": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "CreateLaptopRequest": {
      "type": "object",
      "properties": {
        "laptop": {
          "$ref": "#/definitions/Laptop"
        }
      }
    },
    "CreateLaptopResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "Filter": {
      "type": "object",
      "properties": {
        "maxPriceUsd": {
          "type": "number",
          "format": "double"
        },
        "minCpuCores": {
          "type": "integer",
          "format": "int64"
        },
        "minCpuGhz": {
          "type": "number",
          "format": "double"
        },
        "minRam": {
          "$ref": "#/definitions/Memory"
        }
      }
    },
    "GPU": {
      "type": "object",
      "properties": {
        "brand": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "minGhz": {
          "type": "number",
          "format": "double"
        },
        "maxGhz": {
          "type": "number",
          "format": "double"
        },
        "memory": {
          "$ref": "#/definitions/Memory"
        }
      }
    },
    "ImageInfo": {
      "type": "object",
      "properties": {
        "laptopId": {
          "type": "string"
        },
        "imageType": {
          "type": "string"
        }
      }
    },
    "Keyboard": {
      "type": "object",
      "properties": {
        "layout": {
          "$ref": "#/definitions/KeyboardLayout"
        },
        "backlit": {
          "type": "boolean"
        }
      }
    },
    "KeyboardLayout": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "QWERTY",
        "QWERTZ",
        "AZERTY"
      ],
      "default": "UNKNOWN"
    },
    "Laptop": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "brand": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "cpu": {
          "$ref": "#/definitions/CPU"
        },
        "ram": {
          "$ref": "#/definitions/Memory"
        },
        "gpus": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GPU"
          }
        },
        "storages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Storage"
          }
        },
        "screen": {
          "$ref": "#/definitions/Screen"
        },
        "keyboard": {
          "$ref": "#/definitions/Keyboard"
        },
        "weightKg": {
          "type": "number",
          "format": "double"
        },
        "weightLb": {
          "type": "number",
          "format": "double"
        },
        "priceUsd": {
          "type": "number",
          "format": "double"
        },
        "releaseYear": {
          "type": "integer",
          "format": "int64"
        },
        "updateAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "Memory": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "format": "uint64"
        },
        "unit": {
          "$ref": "#/definitions/MemoryUnit"
        }
      }
    },
    "MemoryUnit": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "BIT",
        "BYTE",
        "KILOBYTE",
        "MEGABYTE",
        "GIGABYTE",
        "TERABYTE"
      ],
      "default": "UNKNOWN"
    },
    "RateLaptopRequest": {
      "type": "object",
      "properties": {
        "laptopId": {
          "type": "string"
        },
        "score": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "RateLaptopResponse": {
      "type": "object",
      "properties": {
        "laptopId": {
          "type": "string"
        },
        "ratedCount": {
          "type": "integer",
          "format": "int64"
        },
        "averageScore": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "Screen": {
      "type": "object",
      "properties": {
        "sizeInch": {
          "type": "number",
          "format": "float"
        },
        "resolution": {
          "$ref": "#/definitions/ScreenResolution"
        },
        "panel": {
          "$ref": "#/definitions/ScreenPanel"
        },
        "multitouch": {
          "type": "boolean"
        }
      }
    },
    "ScreenPanel": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "IPS",
        "OLED"
      ],
      "default": "UNKNOWN"
    },
    "ScreenResolution": {
      "type": "object",
      "properties": {
        "width": {
          "type": "integer",
          "format": "int64"
        },
        "height": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "SearchLaptopResponse": {
      "type": "object",
      "properties": {
        "laptop": {
          "$ref": "#/definitions/Laptop"
        }
      }
    },
    "Storage": {
      "type": "object",
      "properties": {
        "driver": {
          "$ref": "#/definitions/StorageDriver"
        },
        "memory": {
          "$ref": "#/definitions/Memory"
        }
      }
    },
    "StorageDriver": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "HDD",
        "SSD"
      ],
      "default": "UNKNOWN"
    },
    "UploadImageRequest": {
      "type": "object",
      "properties": {
        "info": {
          "$ref": "#/definitions/ImageInfo"
        },
        "chunkData": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "UploadImageResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "size": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}